+++ modules/codemod/lib/v8/softDeprecateDrawer.ts
