+++ modules/codemod/lib/v8/softDeprecatePreviewMenu.ts
